Added windows support for browser ping

+++ octoprint_NanoFactory/__init__.py
@@ -5,6 +5,8 @@ import getpass
 import json
 import requests
 import os
+import platform
+import re
 import subprocess
 import psutil
 import platform
@@ -164,14 +166,26 @@ class NanofactoryPlugin(
         alive = False
 
         if self.pid:
-            try:
-                os.kill(self.pid, 0)
-                alive = True
-                self._plugin_manager.send_plugin_message(
-                    self._identifier, {"browser_status": str(alive)}
-                )
-            except OSError:
-                alive = False
+            if platform.system() == "Windows":
+                out = subprocess.check_output(["tasklist","/fi",f"PID eq {self.pid}"]).strip()
+                # b'INFO: No tasks are running which match the specified criteria.'
+
+                if re.search(b'No tasks', out, re.IGNORECASE):
+                    alive = False
+                else:
+                    alive = True
+            else:
+                try:
+                    os.kill(self.pid, 0)
+                    alive = True
+                except Exception:
+                    alive = False
+
+        self._plugin_manager.send_plugin_message(
+            self._identifier, {"browser_status": str(alive)}
+        )
+
+        self._logger.info(alive)
 
     def is_blueprint_csrf_protected(self):
         return True
